# Improve performance: add inline pragmas