```
Benchmark bench: RUNNING...
benchmarking bigTable/ByteString
time                 1.748 ms   (1.723 ms .. 1.802 ms)         was 2.826 ms   (2.628 ms .. 3.064 ms)
                     0.980 R²   (0.951 R² .. 0.999 R²)
mean                 1.811 ms   (1.769 ms .. 1.888 ms)
std dev              195.1 μs   (88.97 μs .. 297.3 μs)
variance introduced by outliers: 73% (severely inflated)

benchmarking basic/ByteString
time                 5.175 μs   (5.152 μs .. 5.206 μs)         was 7.277 μs   (7.223 μs .. 7.359 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 5.213 μs   (5.178 μs .. 5.297 μs)
std dev              169.4 ns   (76.76 ns .. 311.1 ns)
variance introduced by outliers: 41% (moderately inflated)

benchmarking wideTree/ByteString
time                 1.902 ms   (1.872 ms .. 1.932 ms)         was 2.454 ms   (2.290 ms .. 2.601 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 1.891 ms   (1.884 ms .. 1.902 ms)
std dev              29.83 μs   (18.19 μs .. 44.69 μs)

benchmarking wideTreeEscaping/ByteString
time                 364.5 μs   (362.4 μs .. 366.5 μs)         was 609.8 μs   (510.6 μs .. 814.4 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 364.7 μs   (363.2 μs .. 367.7 μs)
std dev              6.673 μs   (3.712 μs .. 11.91 μs)
variance introduced by outliers: 10% (moderately inflated)

benchmarking deepTree/ByteString
time                 953.4 μs   (943.8 μs .. 966.6 μs)         was 2.117 ms   (2.053 ms .. 2.191 ms)
                     0.998 R²   (0.997 R² .. 1.000 R²)
mean                 953.9 μs   (948.9 μs .. 963.9 μs)
std dev              21.97 μs   (14.04 μs .. 37.39 μs)
variance introduced by outliers: 12% (moderately inflated)

benchmarking manyAttributes/ByteString
time                 11.13 ms   (10.89 ms .. 11.39 ms)         was 11.89 ms   (11.38 ms .. 12.39 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 11.25 ms   (11.06 ms .. 12.30 ms)
std dev              947.2 μs   (88.48 μs .. 2.200 ms)
variance introduced by outliers: 43% (moderately inflated)

benchmarking customAttribute/ByteString
time                 8.488 ms   (8.352 ms .. 8.605 ms)         was 8.712 ms   (8.382 ms .. 9.188 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 8.548 ms   (8.512 ms .. 8.591 ms)
std dev              107.1 μs   (86.73 μs .. 132.4 μs)

Benchmark bench: FINISH
Benchmark bench-io: RUNNING...
benchmarking renderText
time                 2.263 ms   (2.248 ms .. 2.275 ms)         was 3.397 ms   (3.176 ms .. 3.686 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 2.264 ms   (2.256 ms .. 2.277 ms)
std dev              34.21 μs   (25.74 μs .. 49.21 μs)

benchmarking renderTextT Identity
time                 4.999 ms   (3.744 ms .. 6.030 ms)         was 3.060 ms   (2.997 ms .. 3.144 ms)
                     0.824 R²   (0.776 R² .. 0.897 R²)
mean                 2.959 ms   (2.679 ms .. 3.471 ms)
std dev              1.155 ms   (775.0 μs .. 1.698 ms)
variance introduced by outliers: 98% (severely inflated)

benchmarking renderTextT Reader
time                 2.324 ms   (2.273 ms .. 2.368 ms)         was 3.755 ms   (3.662 ms .. 3.860 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 2.523 ms   (2.459 ms .. 2.610 ms)
std dev              250.5 μs   (198.1 μs .. 299.4 μs)
variance introduced by outliers: 67% (severely inflated)

benchmarking renderTextT IO
time                 3.191 ms   (3.161 ms .. 3.230 ms)         was 8.954 ms   (8.620 ms .. 9.612 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 3.282 ms   (3.252 ms .. 3.329 ms)
std dev              120.9 μs   (75.83 μs .. 185.4 μs)
variance introduced by outliers: 20% (moderately inflated)

Benchmark bench-io: FINISH
```

Improvements are clear, resolves #63